requirements.txt
Pillow>=7
numpy
torch>=1.4
torchvision>=0.10
gevent
Flask>=2
requests>=2
psutil